test: global-state react에서의 테스트 추가
